requirements.txt
Flask
Gunicorn
sqlalchemy
python-dotenv
flask-cors
mysqlclient